+++ docker/client/deploy.sh
@@ -2,15 +2,17 @@
 BUILDDIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null && pwd )"
 REPOSITORY="bscdataclay"
 source $BUILDDIR/../../common/config.sh
-source $BUILDDIR/../../common/prepare_docker_builder.sh
+if [ "$SHARE_BUILDERX" = "false" ]; then
+  source $BUILDDIR/../../common/prepare_docker_builder.sh
+fi
 
 # CLIENT 
 pushd $BUILDDIR
 # client will not have execution environemnt in version, like pypi
-echo "************* Building image named $REPOSITORY/client:$CLIENT_TAG *************"
-docker buildx build $DOCKERFILE -t $REPOSITORY/client:$CLIENT_TAG \
+echo "************* Pushing image named $REPOSITORY/client:$CLIENT_TAG (retry $n) *************"
+deploy docker buildx build $DOCKERFILE -t $REPOSITORY/client:$CLIENT_TAG \
          --build-arg VCS_REF=`git rev-parse --short HEAD` \
-         --build-arg BUILD_DATE=`date -u +"%Y-%m-%dT%H:%M:%SZ"` \
+         --build-arg BUILD_DATE=`date -u +"%Y-%m-%dT00:00:00Z"` \
          --build-arg VERSION=$CLIENT_TAG \
 				 --build-arg DATACLAY_DSPYTHON_DOCKER_TAG=$DEFAULT_PY_CLIENT_TAG \
 				 --build-arg DATACLAY_LOGICMODULE_DOCKER_TAG=$DEFAULT_JDK_CLIENT_TAG \
@@ -20,7 +22,8 @@ docker buildx build $DOCKERFILE -t $REPOSITORY/client:$CLIENT_TAG \
 	       --cache-from=type=registry,ref=bscdataclay/client:${CLIENT_TAG}-buildxcache \
 				 --platform $PLATFORMS \
 				 --push .
-echo "************* $REPOSITORY/client:$CLIENT_TAG DONE! *************"
+
+echo "************* $REPOSITORY/client:$CLIENT_TAG IMAGE PUSHED! (in $n retries) *************"
 popd 
 
 if [ "$DEV" = false ] ; then
@@ -31,8 +34,10 @@ else
 fi
 
 # Remove builder
-docker buildx rm $DOCKER_BUILDER
-printMsg " ===== Done! ====="
+if [ "$SHARE_BUILDERX" = "false" ]; then
+  docker buildx rm $DOCKER_BUILDER
+fi
+printMsg " ===== Done! (in $n retries) ====="
