+++ docker/logicmodule/deploy.sh
@@ -1,9 +1,13 @@
 #!/bin/bash
 BUILDDIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null && pwd )"
 REPOSITORY="bscdataclay"
+
 source $BUILDDIR/../../common/config.sh
 if [ -z $EXECUTION_ENVIRONMENT_TAG ]; then echo "ERROR: EXECUTION_ENVIRONMENT_TAG not defined. Aborting"; exit 1; fi
-source $BUILDDIR/../../common/prepare_docker_builder.sh
+
+if [ "$SHARE_BUILDERX" = "false" ]; then
+  source $BUILDDIR/../../common/prepare_docker_builder.sh
+fi
 
 if [ "$PACKAGE_JAR" = "true" ]; then
   # CREATE DATACLAY JAR
@@ -16,10 +20,11 @@ fi
 
 # LOGICMODULE
 pushd $BUILDDIR
-echo "************* Pushing image named $REPOSITORY/logicmodule:$EXECUTION_ENVIRONMENT_TAG *************"
-docker buildx build $DOCKERFILE -t $REPOSITORY/logicmodule:$EXECUTION_ENVIRONMENT_TAG \
+
+echo "************* Pushing image named $REPOSITORY/logicmodule:$EXECUTION_ENVIRONMENT_TAG (retry $n) *************"
+deploy docker buildx build $DOCKERFILE -t $REPOSITORY/logicmodule:$EXECUTION_ENVIRONMENT_TAG \
     --build-arg VCS_REF=`git rev-parse --short HEAD` \
-    --build-arg BUILD_DATE=`date -u +"%Y-%m-%dT%H:%M:%SZ"` \
+    --build-arg BUILD_DATE=`date -u +"%Y-%m-%dT00:00:00Z"` \
     --build-arg VERSION=$EXECUTION_ENVIRONMENT_TAG \
 		--build-arg JDK=$JAVA_VERSION \
 		--build-arg BASE_VERSION=$BASE_VERSION_TAG \
@@ -28,7 +33,7 @@ docker buildx build $DOCKERFILE -t $REPOSITORY/logicmodule:$EXECUTION_ENVIRONMEN
 	  --cache-from=type=registry,ref=bscdataclay/logicmodule:${EXECUTION_ENVIRONMENT_TAG}-buildxcache \
 		--platform $PLATFORMS \
 		--push .
-echo "************* $REPOSITORY/logicmodule:$EXECUTION_ENVIRONMENT_TAG IMAGE PUSHED! *************"
+echo "************* $REPOSITORY/logicmodule:$EXECUTION_ENVIRONMENT_TAG IMAGE PUSHED! (in $n retries) *************"
 popd 
 
 
@@ -50,8 +55,10 @@ fi
 #################################################################################################
 
 # Remove builder
-docker buildx rm $DOCKER_BUILDER
-printMsg " ===== Done! ====="
+if [ "$SHARE_BUILDERX" = "false" ]; then
+  docker buildx rm $DOCKER_BUILDER
+fi
+printMsg " ===== Done! (in $n retries) ===== "
